refactor(storage): isolate batch title admission planning
