+++ crates/kglite/src/graph/mutation/maintain.rs
@@ -5,12 +5,16 @@ use crate::graph::introspection::reporting::{ConnectionOperationReport, NodeOper
 use crate::graph::mutation::batch::{
     BatchProcessor, BatchStats, ConflictHandling, ConnectionBatchProcessor, NodeAction,
 };
+use crate::graph::mutation::batch_title_admission::{
+    node_property_columns, prepare_connection_admission, snapshot_node_titles,
+    ConnectionAdmissionFields,
+};
 use crate::graph::mutation::delete_state::remove_doomed_nodes;
 use crate::graph::mutation::edge_props::{
     intern_edge_props, register_used_edge_property_names, resolve_edge_property_columns,
 };
 use crate::graph::mutation::endpoints::{
-    report_null_id_skips, resolve_endpoints, resolve_pairs, title_column_indices, ResolvedEndpoints,
+    report_null_id_skips, resolve_endpoints, resolve_pairs, ResolvedEndpoints,
 };
 use crate::graph::mutation::rel_constraint_gate::{ConnectionBatchGate, RowFolding};
 use crate::graph::schema::{
@@ -23,51 +27,6 @@ use crate::graph::storage::{GraphRead, GraphWrite};
 use petgraph::graph::{EdgeIndex, NodeIndex};
 use std::collections::{HashMap, HashSet};
 use std::sync::Arc;
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-enum PendingTitleOwner {
-    Existing(NodeIndex),
-    Deferred(String, Value),
-}
-
-struct PendingTitleUpdate {
-    owner: PendingTitleOwner,
-    node_type: String,
-    node_idx: Option<NodeIndex>,
-    id: Value,
-    value: Value,
-}
-
-struct ConnectionTitleInput<'a> {
-    frame: &'a DataFrame,
-    matched: &'a [(usize, NodeIndex, NodeIndex)],
-    deferred: &'a [(usize, Value, Value)],
-    source_type: &'a str,
-    target_type: &'a str,
-    source_id_idx: usize,
-    target_id_idx: usize,
-    titles: &'a ConnectionTitles,
-}
-
-struct ConnectionTitles {
-    source_idx: Option<usize>,
-    target_idx: Option<usize>,
-    source_derived: Option<Vec<Value>>,
-    target_derived: Option<Vec<Value>>,
-}
-
-impl ConnectionTitles {
-    fn value(&self, frame: &DataFrame, source: bool, row: usize) -> Option<Value> {
-        let (index, derived) = if source {
-            (self.source_idx, self.source_derived.as_deref())
-        } else {
-            (self.target_idx, self.target_derived.as_deref())
-        };
-        derived
-            .and_then(|values| values.get(row).cloned())
-            .or_else(|| index.and_then(|column| frame.get_value_by_index(row, column)))
-    }
-}
 
 /// Report returned by `add_properties()`.
 ///
@@ -674,11 +633,6 @@ pub fn add_nodes(
     let should_update_title = node_title_field.is_some();
     let title_field = node_title_field.unwrap_or_else(|| unique_id_field.clone());
     check_data_validity(&df_data, &unique_id_field)?;
-    crate::graph::session::snapshot_dataframe_properties(
-        &graph.graph,
-        &mut df_data,
-        &[&unique_id_field],
-    );
 
     let mut errors = Vec::new();
 
@@ -697,17 +651,13 @@ pub fn add_nodes(
     let title_idx = df_data
         .get_column_index(&title_field)
         .ok_or_else(|| format!("Column '{}' not found", title_field))?;
-    let derived_titles = (title_field == unique_id_field).then(|| {
-        let mut titles: Vec<Value> = (0..df_data.row_count())
-            .map(|row_idx| {
-                df_data
-                    .get_value_by_index(row_idx, title_idx)
-                    .unwrap_or(Value::Null)
-            })
-            .collect();
-        crate::graph::session::snapshot_property_values(&graph.graph, titles.iter_mut());
-        titles
-    });
+    let derived_titles = snapshot_node_titles(
+        graph,
+        &mut df_data,
+        &unique_id_field,
+        &title_field,
+        title_idx,
+    );
 
     // Every refusal happens here, ahead of the first write — see `gate_batch`
     // for why none of the writes below is a safe place for one.
@@ -738,19 +688,7 @@ pub fn add_nodes(
 
     // Property column (name + index) resolved once: no per-row string compares
     // and no per-property HashMap lookup in the loop below.
-    let property_columns: Vec<(String, usize)> = df_data
-        .get_column_names()
-        .into_iter()
-        .filter_map(|col_name| {
-            if col_name != unique_id_field && col_name != title_field {
-                df_data
-                    .get_column_index(&col_name)
-                    .map(|idx| (col_name, idx))
-            } else {
-                None
-            }
-        })
-        .collect();
+    let property_columns = node_property_columns(&df_data, &unique_id_field, &title_field);
 
     // One clock read per call; every row receives the same complete stamp.
     let provenance_stamps = install_type_schema(graph, &node_type, &property_columns);
@@ -1111,56 +1049,20 @@ pub(crate) fn add_connections_with_initial_load(
     // A source/target type that doesn't exist yet is not an error: an edge to a
     // missing endpoint vivifies a stub node, which registers the type (Pass B).
 
-    let source_id_idx = df_data
-        .get_column_index(&source_id_field)
-        .ok_or_else(|| format!("Source ID column '{}' not found", source_id_field))?;
-    let target_id_idx = df_data
-        .get_column_index(&target_id_field)
-        .ok_or_else(|| format!("Target ID column '{}' not found", target_id_field))?;
-
-    let (source_title_idx, target_title_idx) = title_column_indices(
-        &df_data,
-        source_title_field.as_deref(),
-        target_title_field.as_deref(),
-    );
-
-    crate::graph::session::snapshot_dataframe_properties(
-        &graph.graph,
+    // Resolve and gate every row before the mutating passes below. The helper
+    // also snapshots stored values while preserving endpoint identity cells.
+    let (resolved, titles) = prepare_connection_admission(
+        graph,
         &mut df_data,
-        &[source_id_field.as_str(), target_id_field.as_str()],
-    );
-    let derive_titles = |column: usize| {
-        let mut values: Vec<Value> = (0..df_data.row_count())
-            .map(|row| {
-                df_data
-                    .get_value_by_index(row, column)
-                    .unwrap_or(Value::Null)
-            })
-            .collect();
-        crate::graph::session::snapshot_property_values(&graph.graph, values.iter_mut());
-        values
-    };
-    let titles = ConnectionTitles {
-        source_idx: source_title_idx,
-        target_idx: target_title_idx,
-        source_derived: source_title_idx
-            .filter(|index| *index == source_id_idx || *index == target_id_idx)
-            .map(derive_titles),
-        target_derived: target_title_idx
-            .filter(|index| *index == source_id_idx || *index == target_id_idx)
-            .map(derive_titles),
-    };
-
-    // Endpoint resolution is its own pass over the frame, ahead of every
-    // mutation, because the id-index probe borrows `graph.id_indices` while
-    // the batch needs `&mut graph`. Splitting it that way is what lets the
-    // probe read the index *in place*: the per-call materialized
-    // `id -> NodeIndex` map it replaces cost one insert per node of the whole
-    // endpoint type and was 53% of a property-free `add_connections` at 100k
-    // nodes / 24k edges (samply, 2026-08-15), growing with the graph while
-    // the row count stayed fixed. It is snapshot-equivalent to the old code:
-    // that map was also built before the first mutation, and nothing in the
-    // mutating passes below moves an existing node's id.
+        ConnectionAdmissionFields {
+            source_type: &source_type,
+            source_id: &source_id_field,
+            source_title: source_title_field.as_deref(),
+            target_type: &target_type,
+            target_id: &target_id_field,
+            target_title: target_title_field.as_deref(),
+        },
+    )?;
     let ResolvedEndpoints {
         matched,
         deferred,
@@ -1168,27 +1070,7 @@ pub(crate) fn add_connections_with_initial_load(
         missing_targets,
         null_source_rows: skipped_null_source,
         null_target_rows: skipped_null_target,
-    } = resolve_endpoints(
-        graph,
-        &df_data,
-        &source_type,
-        &target_type,
-        source_id_idx,
-        target_id_idx,
-    )?;
-    gate_connection_title_updates(
-        graph,
-        ConnectionTitleInput {
-            frame: &df_data,
-            matched: &matched,
-            deferred: &deferred,
-            source_type: &source_type,
-            target_type: &target_type,
-            source_id_idx,
-            target_id_idx,
-            titles: &titles,
-        },
-    )?;
+    } = resolved;
     let mut batch = ConnectionBatchProcessor::new(df_data.row_count());
     batch.set_conflict_mode(conflict_mode);
     // Skip edge existence checks when this load owns every edge of the type.
@@ -1235,16 +1117,13 @@ pub(crate) fn add_connections_with_initial_load(
 
     // Pass A — connect the rows whose endpoints both exist (resolved above).
     for (row_idx, source_idx, target_idx) in matched {
-        update_node_titles(
+        titles.apply(
             graph,
-            &source_type,
-            &target_type,
-            source_idx,
-            target_idx,
+            (&source_type, &target_type),
+            (source_idx, target_idx),
             row_idx,
-            &titles,
             &df_data,
-        )?;
+        );
         if let Err(e) = batch.add_connection(
             source_idx,
             target_idx,
@@ -1282,16 +1161,13 @@ pub(crate) fn add_connections_with_initial_load(
                     continue;
                 }
             };
-            update_node_titles(
+            titles.apply(
                 graph,
-                &source_type,
-                &target_type,
-                source_idx,
-                target_idx,
+                (&source_type, &target_type),
+                (source_idx, target_idx),
                 row_idx,
-                &titles,
                 &df_data,
-            )?;
+            );
             if let Err(e) = batch.add_connection(
                 source_idx,
                 target_idx,
@@ -1305,12 +1181,7 @@ pub(crate) fn add_connections_with_initial_load(
         }
     }
 
-    if titles.source_idx.is_some() {
-        graph.refresh_indexes_for_type(&source_type);
-    }
-    if titles.target_idx.is_some() && target_type != source_type {
-        graph.refresh_indexes_for_type(&target_type);
-    }
+    titles.refresh_indexes(graph, &source_type, &target_type);
 
     report_null_id_skips(
         &mut errors,
@@ -1929,196 +1800,6 @@ pub fn purge_provisional_nodes(graph: &mut DirGraph) -> (usize, usize) {
         }
     }
     detach_delete_nodes(graph, &to_delete)
-}
-
-fn update_node_titles(
-    graph: &mut DirGraph,
-    source_type: &str,
-    target_type: &str,
-    source_idx: NodeIndex,
-    target_idx: NodeIndex,
-    row_idx: usize,
-    titles: &ConnectionTitles,
-    df_data: &DataFrame,
-) -> Result<(), String> {
-    if titles.source_idx.is_some() {
-        if let Some(title) = titles.value(df_data, true, row_idx) {
-            GraphWrite::set_node_title(&mut graph.graph, source_idx, title);
-            crate::graph::index_freshness::write_hooks::note_property_written(
-                graph,
-                source_idx,
-                source_type,
-                Some("title"),
-            );
-        }
-    }
-    if titles.target_idx.is_some() {
-        if let Some(title) = titles.value(df_data, false, row_idx) {
-            GraphWrite::set_node_title(&mut graph.graph, target_idx, title);
-            crate::graph::index_freshness::write_hooks::note_property_written(
-                graph,
-                target_idx,
-                target_type,
-                Some("title"),
-            );
-        }
-    }
-    Ok(())
-}
-
-fn gate_connection_title_updates(
-    graph: &mut DirGraph,
-    input: ConnectionTitleInput<'_>,
-) -> Result<(), String> {
-    if input.titles.source_idx.is_none() && input.titles.target_idx.is_none() {
-        return Ok(());
-    }
-
-    let updates = collect_connection_title_updates(&input);
-    let mut batch_claims = HashMap::new();
-    for update in updates {
-        if let Some(shape) = graph.shape_for(&update.node_type, "title") {
-            shape.check("title", &update.value)?;
-        }
-        gate_pending_title_update(graph, &update, &mut batch_claims)?;
-    }
-    Ok(())
-}
-
-fn collect_connection_title_updates(input: &ConnectionTitleInput<'_>) -> Vec<PendingTitleUpdate> {
-    let mut updates = Vec::<PendingTitleUpdate>::new();
-    let mut positions = HashMap::<PendingTitleOwner, usize>::new();
-    let mut remember = |update: PendingTitleUpdate| {
-        if let Some(position) = positions.get(&update.owner).copied() {
-            updates[position] = update;
-        } else {
-            positions.insert(update.owner.clone(), updates.len());
-            updates.push(update);
-        }
-    };
-    let value_at = |row: usize, column: usize| {
-        input
-            .frame
-            .get_value_by_index(row, column)
-            .unwrap_or(Value::Null)
-    };
-
-    for (row, source, target) in input.matched {
-        if input.titles.source_idx.is_some() {
-            remember(PendingTitleUpdate {
-                owner: PendingTitleOwner::Existing(*source),
-                node_type: input.source_type.to_string(),
-                node_idx: Some(*source),
-                id: value_at(*row, input.source_id_idx),
-                value: input
-                    .titles
-                    .value(input.frame, true, *row)
-                    .unwrap_or(Value::Null),
-            });
-        }
-        if input.titles.target_idx.is_some() {
-            remember(PendingTitleUpdate {
-                owner: PendingTitleOwner::Existing(*target),
-                node_type: input.target_type.to_string(),
-                node_idx: Some(*target),
-                id: value_at(*row, input.target_id_idx),
-                value: input
-                    .titles
-                    .value(input.frame, false, *row)
-                    .unwrap_or(Value::Null),
-            });
-        }
-    }
-    for (row, source_id, target_id) in input.deferred {
-        if input.titles.source_idx.is_some() {
-            remember(PendingTitleUpdate {
-                owner: PendingTitleOwner::Deferred(
-                    input.source_type.to_string(),
-                    source_id.clone(),
-                ),
-                node_type: input.source_type.to_string(),
-                node_idx: None,
-                id: source_id.clone(),
-                value: input
-                    .titles
-                    .value(input.frame, true, *row)
-                    .unwrap_or(Value::Null),
-            });
-        }
-        if input.titles.target_idx.is_some() {
-            remember(PendingTitleUpdate {
-                owner: PendingTitleOwner::Deferred(
-                    input.target_type.to_string(),
-                    target_id.clone(),
-                ),
-                node_type: input.target_type.to_string(),
-                node_idx: None,
-                id: target_id.clone(),
-                value: input
-                    .titles
-                    .value(input.frame, false, *row)
-                    .unwrap_or(Value::Null),
-            });
-        }
-    }
-
-    updates
-}
-
-fn gate_pending_title_update(
-    graph: &mut DirGraph,
-    update: &PendingTitleUpdate,
-    batch_claims: &mut HashMap<(UniqueConstraintKey, CompositeValue), PendingTitleOwner>,
-) -> Result<(), String> {
-    let claims = if let Some(node_idx) = update.node_idx {
-        graph
-            .plan_property_write(&update.node_type, node_idx, "title", Some(&update.value))
-            .map(|plan| plan.claim)
-            .map_err(|violation| violation.to_string())?
-    } else {
-        let typed = graph.check_property_types(&update.node_type, |property| {
-            deferred_title_field(graph, update, property)
-        });
-        if let Err(violation) = typed {
-            return Err(graph.record_constraint_violation(*violation));
-        }
-        let required = graph.check_required_fields(&update.node_type, |property| {
-            deferred_title_field(graph, update, property)
-        });
-        if let Err(violation) = required {
-            return Err(graph.record_constraint_violation(*violation));
-        }
-        let claims = graph.unique_claims(&update.node_type, |property| {
-            deferred_title_field(graph, update, property)
-        });
-        graph
-            .check_unique_claims(&claims, None)
-            .map_err(|violation| graph.record_constraint_violation(*violation))?;
-        claims
-    };
-    for claim in claims {
-        let key = (claim.key.clone(), claim.value.clone());
-        if let Some(owner) = batch_claims.insert(key, update.owner.clone()) {
-            if owner != update.owner {
-                let violation = graph.unique_batch_conflict(&claim);
-                return Err(graph.record_constraint_violation(violation));
-            }
-        }
-    }
-    Ok(())
-}
-
-fn deferred_title_field(
-    graph: &DirGraph,
-    update: &PendingTitleUpdate,
-    property: &str,
-) -> Option<Value> {
-    match graph.resolve_alias(&update.node_type, property) {
-        "id" => Some(update.id.clone()),
-        "title" => (!matches!(update.value, Value::Null)).then(|| update.value.clone()),
-        PROVISIONAL_KEY => Some(Value::Boolean(true)),
-        _ => None,
-    }
 }
 
 fn update_schema_node(
