+++ tests/api-baselines/source-quality.json
@@ -325,8 +325,8 @@
     {
       "path": "crates/kglite/src/graph/mutation/maintain.rs",
       "qualified_name": "crate::add_connections_with_initial_load",
-      "lines": 249,
-      "branches": 27,
+      "lines": 231,
+      "branches": 23,
       "nesting": 3
     },
     {
